Moved the files in code-customization folder
