+++ packages/core/src/index.ts
@@ -68,6 +68,7 @@ export * from './tools/code-customization-setup/DeleteGroups.js';
 export * from './tools/code-customization-setup/UpdateGroups.js';
 export * from './tools/code-customization-setup/ListRepoLinks.js';
 export * from './tools/code-customization-setup/GetRepoLinks.js';
+export * from './tools/code-customization-setup/CreateRepoLinks.js';
 
 // Export telemetry functions
 export * from './telemetry/index.js';
